requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
charset-normalizer==2.1.1
discord==1.0.1
discord.py==1.6.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
typing_extensions==4.4.0
yarl==1.8.1
xmltodict==0.13.0